Initial commit
This adds the interactive examples for the first chapter of the book. It
contains the initial implementation of the algorithms in the CompGeom
module needed by the examples as well as supporting code.

As these examples mostly pertain to points in a plane, some plugins for
d3 has been extracted for convenience and reusability:

- d3.vertex
- d3.selectionBox

See comments in ./scripts/d3_plugins for further information.

CSS is compiled using sass and js bundles with browserify, which both
have specified run commands in package.json.

Tests are run using the Jasmine test harness and lives in ./test.